requirements.txt

fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
